fix GroveDB persistence
