+++ grovedb/src/lib.rs
@@ -1,4 +1,8 @@
 #![feature(trivial_bounds)]
+mod subtree;
+#[cfg(test)]
+mod tests;
+
 use std::{
     collections::{HashMap, HashSet},
     path::Path,
@@ -8,9 +12,10 @@ use std::{
 use merk::{self, rocksdb, Merk};
 use rs_merkle::{algorithms::Sha256, MerkleTree};
 use subtree::Element;
-mod subtree;
 
 const MAX_REFERENCE_HOPS: usize = 10;
+
+const SUBTRESS_SERIALIZED_KEY: &[u8] = b"subtreesSerialized";
 
 // Root tree has hardcoded leafs; each of them is `pub` to be easily used in
 // `path` arg
@@ -25,7 +30,7 @@ pub const DATA_CONTRACTS_TREE_KEY: &[u8] = b"dataContracts";
 #[derive(Debug, thiserror::Error)]
 pub enum Error {
     #[error("rocksdb error")]
-    RocksDBError(#[from] rocksdb::Error),
+    RocksDBError(#[from] merk::rocksdb::Error),
     #[error("unable to open Merk db")]
     MerkError(merk::Error),
     #[error("invalid path")]
@@ -57,17 +62,27 @@ impl GroveDb {
             path,
             merk::column_families(),
         )?);
+
         let mut subtrees = HashMap::new();
-        // TODO: this will work only for a fresh RocksDB and it cannot restore any
-        // other subtree than these constants at any level!
-        for subtree_path in [
-            COMMON_TREE_KEY,
-            IDENTITIES_TREE_KEY,
-            PUBLIC_KEYS_TO_IDENTITY_IDS_TREE_KEY,
-            DATA_CONTRACTS_TREE_KEY,
-        ] {
-            let subtree_merk = Merk::open(db.clone(), subtree_path.to_vec())?;
-            subtrees.insert(subtree_path.to_vec(), subtree_merk);
+        if let Some(prefixes_compressed) = db.get(SUBTRESS_SERIALIZED_KEY)? {
+            let subtrees_prefixes: Vec<Vec<u8>> = bincode::deserialize(&prefixes_compressed)?;
+            for prefix in subtrees_prefixes {
+                let subtree_merk = Merk::open(db.clone(), prefix.to_vec())?;
+                subtrees.insert(prefix.to_vec(), subtree_merk);
+            }
+        } else {
+            // TODO: this will work only for a fresh RocksDB and it cannot restore any
+            // other subtree than these constants at any level!
+            for subtree_path in [
+                COMMON_TREE_KEY,
+                IDENTITIES_TREE_KEY,
+                PUBLIC_KEYS_TO_IDENTITY_IDS_TREE_KEY,
+                DATA_CONTRACTS_TREE_KEY,
+            ] {
+                let subtree_merk = Merk::open(db.clone(), subtree_path.to_vec())?;
+                subtrees.insert(subtree_path.to_vec(), subtree_merk);
+            }
+            Self::store_subtrees_prefixes(&subtrees, &db)?;
         }
 
         Ok(GroveDb {
@@ -75,6 +90,14 @@ impl GroveDb {
             db: db.clone(),
             subtrees,
         })
+    }
+
+    fn store_subtrees_prefixes(
+        subtrees: &HashMap<Vec<u8>, Merk>,
+        db: &rocksdb::DB,
+    ) -> Result<(), Error> {
+        let prefixes: Vec<Vec<u8>> = subtrees.keys().map(|x| x.clone()).collect();
+        Ok(db.put(SUBTRESS_SERIALIZED_KEY, bincode::serialize(&prefixes)?)?)
     }
 
     // TODO: evntually there should be no hardcoded root tree structure
@@ -135,6 +158,7 @@ impl GroveDb {
                     element.insert(&mut merk, key)?;
                     self.propagate_changes(path)?;
                 }
+                Self::store_subtrees_prefixes(&self.subtrees, &self.db)?;
             }
             _ => {
                 // If path is empty that means there is an attempt to insert something into a
@@ -246,174 +270,5 @@ impl GroveDb {
             res.extend_from_slice(k);
         }
         res
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use tempdir::TempDir;
-
-    use super::*;
-
-    #[test]
-    fn test_init() {
-        let tmp_dir = TempDir::new("db").unwrap();
-        GroveDb::open(tmp_dir).expect("empty tree is ok");
-    }
-
-    #[test]
-    fn test_insert_value_to_merk() {
-        let tmp_dir = TempDir::new("db").unwrap();
-        let mut db = GroveDb::open(tmp_dir).unwrap();
-        let element = Element::Item(b"ayy".to_vec());
-        db.insert(&[COMMON_TREE_KEY], b"key".to_vec(), element.clone())
-            .expect("successful insert");
-        assert_eq!(
-            db.get(&[COMMON_TREE_KEY], b"key").expect("succesful get"),
-            element
-        );
-    }
-
-    #[test]
-    fn test_insert_value_to_subtree() {
-        let tmp_dir = TempDir::new("db").unwrap();
-        let mut db = GroveDb::open(tmp_dir).unwrap();
-        let element = Element::Item(b"ayy".to_vec());
-
-        // Insert a subtree first
-        db.insert(&[COMMON_TREE_KEY], b"key1".to_vec(), Element::empty_tree())
-            .expect("successful subtree insert");
-        // Insert an element into subtree
-        db.insert(
-            &[COMMON_TREE_KEY, b"key1"],
-            b"key2".to_vec(),
-            element.clone(),
-        )
-        .expect("successful value insert");
-        assert_eq!(
-            db.get(&[COMMON_TREE_KEY, b"key1"], b"key2")
-                .expect("succesful get"),
-            element
-        );
-    }
-
-    #[test]
-    fn test_changes_propagated() {
-        let tmp_dir = TempDir::new("db").unwrap();
-        let mut db = GroveDb::open(tmp_dir).unwrap();
-        let old_hash = db.root_tree.root();
-        let element = Element::Item(b"ayy".to_vec());
-
-        // Insert some nested subtrees
-        db.insert(&[COMMON_TREE_KEY], b"key1".to_vec(), Element::empty_tree())
-            .expect("successful subtree 1 insert");
-        db.insert(
-            &[COMMON_TREE_KEY, b"key1"],
-            b"key2".to_vec(),
-            Element::empty_tree(),
-        )
-        .expect("successful subtree 2 insert");
-        // Insert an element into subtree
-        db.insert(
-            &[COMMON_TREE_KEY, b"key1", b"key2"],
-            b"key3".to_vec(),
-            element.clone(),
-        )
-        .expect("successful value insert");
-        assert_eq!(
-            db.get(&[COMMON_TREE_KEY, b"key1", b"key2"], b"key3")
-                .expect("succesful get"),
-            element
-        );
-        assert_ne!(old_hash, db.root_tree.root());
-    }
-
-    #[test]
-    fn test_follow_references() {
-        let tmp_dir = TempDir::new("db").unwrap();
-        let mut db = GroveDb::open(tmp_dir).unwrap();
-        let element = Element::Item(b"ayy".to_vec());
-
-        // Insert a reference
-        db.insert(
-            &[COMMON_TREE_KEY],
-            b"reference_key".to_vec(),
-            Element::Reference(vec![
-                COMMON_TREE_KEY.to_vec(),
-                b"key2".to_vec(),
-                b"key3".to_vec(),
-            ]),
-        )
-        .expect("successful reference insert");
-
-        // Insert an item to refer to
-        db.insert(&[COMMON_TREE_KEY], b"key2".to_vec(), Element::empty_tree())
-            .expect("successful subtree 1 insert");
-        db.insert(
-            &[COMMON_TREE_KEY, b"key2"],
-            b"key3".to_vec(),
-            element.clone(),
-        )
-        .expect("successful value insert");
-        assert_eq!(
-            db.get(&[COMMON_TREE_KEY], b"reference_key")
-                .expect("succesful get"),
-            element
-        );
-    }
-
-    #[test]
-    fn test_cyclic_references() {
-        let tmp_dir = TempDir::new("db").unwrap();
-        let mut db = GroveDb::open(tmp_dir).unwrap();
-
-        db.insert(
-            &[COMMON_TREE_KEY],
-            b"reference_key_1".to_vec(),
-            Element::Reference(vec![COMMON_TREE_KEY.to_vec(), b"reference_key_2".to_vec()]),
-        )
-        .expect("successful reference 1 insert");
-
-        db.insert(
-            &[COMMON_TREE_KEY],
-            b"reference_key_2".to_vec(),
-            Element::Reference(vec![COMMON_TREE_KEY.to_vec(), b"reference_key_1".to_vec()]),
-        )
-        .expect("successful reference 2 insert");
-
-        assert!(matches!(
-            db.get(&[COMMON_TREE_KEY], b"reference_key_1").unwrap_err(),
-            Error::CyclicReference
-        ));
-    }
-
-    #[test]
-    fn test_too_many_indirections() {
-        let tmp_dir = TempDir::new("db").unwrap();
-        let mut db = GroveDb::open(tmp_dir).unwrap();
-
-        let keygen = |idx| format!("key{}", idx).bytes().collect::<Vec<u8>>();
-
-        db.insert(
-            &[COMMON_TREE_KEY],
-            b"key0".to_vec(),
-            Element::Item(b"oops".to_vec()),
-        )
-        .expect("successful item insert");
-
-        for i in 1..=(MAX_REFERENCE_HOPS + 1) {
-            db.insert(
-                &[COMMON_TREE_KEY],
-                keygen(i),
-                Element::Reference(vec![COMMON_TREE_KEY.to_vec(), keygen(i - 1)]),
-            )
-            .expect("successful reference insert");
-        }
-
-        assert!(matches!(
-            db.get(&[COMMON_TREE_KEY], &keygen(MAX_REFERENCE_HOPS + 1))
-                .unwrap_err(),
-            Error::ReferenceLimit
-        ));
     }
 }
